+++ altium/src/draw/mod.rs
@@ -1,15 +1,19 @@
+//! Tools related to drawing objects
+
 pub(crate) mod canvas;
 mod svg;
 
 pub use canvas::{Canvas, DrawImage, DrawLine, DrawPolygon, DrawRectangle, DrawText};
 
 pub use self::svg::SvgCtx;
-pub use crate::common::{Color, Location, PosHoriz, PosVert};
+pub use crate::common::{Location, PosHoriz, PosVert, Rgb};
 
+/// Implementors of this trait can draw themselves to a canvas.
 pub trait Draw {
+    /// Additional context needed to draw
     type Context<'a>;
 
-    /// Draw this element to a SVG and return the new SVG
+    /// Draw this element to a [`Canvas`].
     ///
     /// This has a defualt implementation that does nothing for easier
     /// reusability