+++ altium/src/lib.rs
@@ -1,3 +1,4 @@
+#![doc = include_str!("../README.md")]
 //! Tools for to reading and writing files produced by Altium Designer.
 //!
 //! It is very early in development, so please expect surprises if you are using
@@ -31,7 +32,6 @@ mod parse;
 
 #[doc(hidden)]
 pub mod __private;
-#[doc(hidden)] // docs are unfinished
 pub mod draw;
 pub mod dwf;
 pub mod error;
@@ -41,7 +41,7 @@ pub mod prj;
 pub mod sch;
 
 #[doc(inline)]
-pub use common::UniqueId;
+pub use common::{Location, Rgb, Rotation90, UniqueId, Visibility};
 #[doc(inline)]
 pub use error::{Error, ErrorKind, Result};
 #[doc(inline)]
@@ -51,8 +51,6 @@ pub use prj::PrjPcb;
 #[doc(inline)]
 pub use sch::{SchDoc, SchLib};
 
-#[cfg(doctest)]
-mod readme_tests {
-    #[doc = include_str!("../../README.md")]
-    struct MainReadMe;
+mod sealed {
+    pub trait Sealed {}
 }